Apply local phong lighting to MapLayer and Grid3D layer as already done in TriangleLayer.    - This algorithm is a modification of Luma's version in that it accounts for two sided lighting.    - also rename privatLayer.ts to privatGrid3dLayer.ts to align with naming convention
