+++ typescript/packages/subsurface-viewer/src/layers/grid3d/fragment.fs.glsl.ts
@@ -56,8 +56,8 @@ void main(void) {
    }
 
    // Use normal lighting. This has no effect if "material" property is not set.
-   vec3 lightColor = lighting_getLightColor(color.rgb, cameraPosition, position_commonspace.xyz, normal);
-   gl_FragColor = vec4(lightColor, 1.0);   
+   vec3 lightColor = getPhongLightColor(color.rgb, cameraPosition, position_commonspace.xyz, normal);
+   gl_FragColor = vec4(lightColor, 1.0);
    DECKGL_FILTER_COLOR(gl_FragColor, geometry);
 }
 `;
