+++ typescript/packages/subsurface-viewer/src/layers/grid3d/privateGrid3dLayer.ts
